pyproject.toml
[tool.poetry]
name = "city-issues-bot"
version = "1.0.0"
description = "Telegram bot for urban issues"
authors = ["AI Assistant <gemini@google.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.4.1"
aiosqlite = "^0.19.0"
python-dotenv = "^1.0.1"
apscheduler = "^3.10.4"
pytz = "^2024.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
ruff = "^0.2.2"
mypy = "^1.8.0"
